heartbeat/requirements.txt
schedule>=1.2.0
requests>=2.31.0
pyyaml>=6.0
